Refactor: Remove unused exception classes and restructure package hierarchy
