--- src/main/java/info/mackiewicz/bankapp/presentation/dashboard/exception/UsernameLengthException.java
@@ -1,17 +1,0 @@
-package info.mackiewicz.bankapp.presentation.dashboard.exception;
-
-import info.mackiewicz.bankapp.shared.core.error.ErrorCode;
-
-public class UsernameLengthException extends CredentialsBaseException {
-
-    private static final ErrorCode ERROR_CODE = ErrorCode.USERNAME_LENGTH;
-
-    public UsernameLengthException(String message) {
-        super(message, ERROR_CODE);
-    }
-
-    public UsernameLengthException(String message, Throwable cause) {
-        super(message, cause, ERROR_CODE);
-    }
-
-}
