+++ src/main/java/info/mackiewicz/bankapp/shared/config/SecurityConfig.java
@@ -1,5 +1,6 @@
 package info.mackiewicz.bankapp.shared.config;
 
+import info.mackiewicz.bankapp.shared.service.CustomUserDetailsService;
 import info.mackiewicz.bankapp.user.service.AdminUserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
