+++ src/test/java/info/mackiewicz/bankapp/presentation/dashboard/service/PasswordValidationServiceTest.java
@@ -1,9 +1,8 @@
-package info.mackiewicz.bankapp.security.service;
+package info.mackiewicz.bankapp.presentation.dashboard.service;
 
 import info.mackiewicz.bankapp.presentation.dashboard.exception.InvalidPasswordException;
 import info.mackiewicz.bankapp.presentation.dashboard.exception.PasswordSameException;
 import info.mackiewicz.bankapp.presentation.dashboard.exception.PasswordsMismatchException;
-import info.mackiewicz.bankapp.presentation.dashboard.service.PasswordValidationService;
 import info.mackiewicz.bankapp.system.security.password.PasswordService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
