+++ src/test/java/info/mackiewicz/bankapp/presentation/dashboard/service/UsernameValidationServiceTest.java
@@ -1,9 +1,8 @@
-package info.mackiewicz.bankapp.security.service;
+package info.mackiewicz.bankapp.presentation.dashboard.service;
 
 import info.mackiewicz.bankapp.presentation.dashboard.exception.ForbiddenUsernameException;
 import info.mackiewicz.bankapp.presentation.dashboard.exception.UsernameAlreadyTakenException;
 import info.mackiewicz.bankapp.presentation.dashboard.exception.UsernameSameException;
-import info.mackiewicz.bankapp.presentation.dashboard.service.UsernameValidationService;
 import info.mackiewicz.bankapp.user.service.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
